Refactor highlighting
